+++ src/main/java/com/cook4you/model/Offer.java
@@ -1,4 +1,4 @@
-package model;
+package com.cook4you.model;
 
 /**
  * Created by Paweł Proc on 2016-11-29.
